+++ cms/home/models/__init__.py
@@ -1,2 +1,2 @@
-from .homepage import HomePage, HomePageRelatedLink, UKHSARootPage
-from .landingpage import LandingPage
+from .home_page import HomePage, HomePageRelatedLink, UKHSARootPage
+from .landing_page import LandingPage
